feat: add resume page metadata, fix skeleton to match layout
- Convert resume page to server/client split for metadata support
- Add SEO metadata to resume page
- Replace generic skeleton with accurate resume preview skeleton
- Use skeleton instead of spinner during loading
